Feature gunjan

+++ public/js/deleteart.js
@@ -1,17 +1,18 @@
-async function deleteAnArt() {
-  const id = window.location.toString().split("/")[
-    window.location.toString().split("/").length - 1
-  ];
+async function deleteAnArt(event) {
+  if (event.target.hasAttribute("data-id")) {
+    const id = event.target.getAttribute("data-id");
+    // console.log("@@@@@@@@@@@@@@@@@@@@@@@", id);
+    const response = await fetch(`/api/arts/${id}`, {
+      method: "DELETE",
+    });
 
-  const response = await fetch("/api/arts/${id}", {
-    method: "DELETE",
-  });
-
-  if (response.ok) {
-    document.location.replace("/dashboard/");
-  } else {
-    alert("Art Not Deleted");
+    if (response.ok) {
+      document.location.replace("/dashboard");
+    } else {
+      alert("Failed to delete art");
+    }
   }
 }
 
-document.querySelector(".artdelete").addEventListener("click", deleteAnArt);
+// document.querySelector("#deleteart").addEventListener("click", deleteAnArt);
+document.querySelector("#allartsbyuser").addEventListener("click", deleteAnArt);
